Moved TooManyPins modal to separate component. Set max pins to 8. [RCORE-600]
